requirements.txt
python-dotenv==1.2.2
spoonacular==3.0
